+++ BookingCoffeeTable/src/main/webapp/views/admin/table-management.jsp
@@ -1,4 +1,4 @@
-<%@include file="/common/taglib.jsp"%>
+<%@include file="/common/taglib.jsp" %>
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <!DOCTYPE html>
 <html lang="en">
@@ -22,6 +22,7 @@
     <link rel="stylesheet" href="<c:url value="/views/template/mdb/plugins/css/all.min.css"/> ">
     <!-- Custom styles -->
     <link rel="stylesheet" href="<c:url value="/views/admin/assets/css/home.css"/> ">
+    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/sweetalert2@11">
     <style></style>
 </head>
 
@@ -98,7 +99,8 @@
                             <a href="/admin/table-management/edit?id=${table.id}" class="btn btn-primary btn-floating">
                                 <i class="far fa-pen-to-square"></i>
                             </a>
-                            <button type="button" class="btn btn-danger btn-floating" onclick="deleteTable(${table.id})">
+                            <button type="button" class="btn btn-danger btn-floating"
+                                    onclick="deleteTable(${table.id})">
                                 <i class="far fa-trash-can"></i>
                             </button>
                         </td>
@@ -121,23 +123,52 @@
 
 <!-- MDB ESSENTIAL -->
 <script src="<c:url value="/views/template/assets/js/jquery-2.1.0.min.js"/> "></script>
-<script>
-    function deleteTable(id){
 
-        $.ajax({
-            url: '/api/admin/table',
-            contentType: "application/json",
-            type: 'DELETE',
-            data: JSON.stringify({
-                id: id
-            }),
-            success: function (data) {
-                if (data) {
-                    alert('Xóa thành công');
-                    location.reload();
-                } else {
-                    alert('Xóa thất bại');
-                }
+<script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+<script>
+    function deleteTable(id) {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "Bạn sẽ không thể quay lại!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                // User confirmed, proceed with the AJAX request
+                $.ajax({
+                    url: '/api/admin/table',
+                    contentType: 'application/json',
+                    type: 'DELETE',
+                    data: JSON.stringify({
+                        id: id
+                    }),
+                    success: function (data) {
+                        if (data) {
+                            Swal.fire({
+                                title: 'Deleted!',
+                                text: 'Đã xóa.',
+                                icon: 'success'
+                            }).then(() => {
+                                location.reload();
+                            });
+                        } else {
+                            Swal.fire({
+                                icon: "error",
+                                title: "Oops...",
+                                text: "Đã có chuyện gì đó xảy ra!",
+                            });
+                        }
+                    }, error: function (e) {
+                        Swal.fire({
+                            icon: "error",
+                            title: "Oops...",
+                            text: "Dữ liệu nằm trong dữ liệu khác!",
+                        });
+                    }
+                });
             }
         });
     }
